requirements.txt
python-telegram-bot>=21.5
sqlalchemy
aiohttp>=3.10.0
requests>=2.32.0
pytz
python-dotenv
psycopg2-binary
Pillow
PyMuPDF
python-docx
google-generativeai>=0.8.5
